1. product list 페이지 완료 2. 입력/수정페이지 jsp 수정중(지역 js 개발 필요, 수정 update, insert 구분 개발 필요)-박미선

+++ GLtest01/src/main/java/com/gl/master/model/product/ProductVo.java
@@ -17,8 +17,8 @@ public class ProductVo {
 	private Date startd;
 	private String img;
 	private String thumb;
-	private int discount;
-	private int point;
+	private double discount;
+	private double point;
 	private int prio;
 	private int event;
 	public ProductVo() {
@@ -132,16 +132,16 @@ public class ProductVo {
 	public void setThumb(String thumb) {
 		this.thumb = thumb;
 	}
-	public int getDiscount() {
+	public double getDiscount() {
 		return discount;
 	}
-	public void setDiscount(int discount) {
+	public void setDiscount(double discount) {
 		this.discount = discount;
 	}
-	public int getPoint() {
+	public double getPoint() {
 		return point;
 	}
-	public void setPoint(int point) {
+	public void setPoint(double point) {
 		this.point = point;
 	}
 	public int getPrio() {
